requirements.txt
langchain~=0.0.276
streamlit~=1.26.0
python-dotenv~=1.0.0
cassandra-driver~=3.28.0
openai~=0.27.9
cassio~=0.1.1

